+++ tests/dont_run_these_in_parallel_test.rb
@@ -20,4 +20,10 @@ class DontRunTheseInParallelTest < Minitest::Test
       assert_strings_appear_in_this_order result.stdout, ["TA#test_2", other]
     end
   end
+
+  def test_not_running_parallel_specified_by_superclass
+    result = TLDRunner.should_succeed "dont_run_these_in_parallel_superclasses.rb"
+
+    assert_includes result.stdout, "2 test methods"
+  end
 end