+++ tests/run_these_together_test.rb
@@ -6,4 +6,10 @@ class RunTheseTogetherTest < Minitest::Test
 
     assert_includes result.stdout, "6 test methods"
   end
+
+  def test_running_these_together_specified_by_superclass
+    result = TLDRunner.should_succeed "run_these_together_superclasses.rb"
+
+    assert_includes result.stdout, "2 test methods"
+  end
 end